+++ services/party/internal/handler/update_party.go
@@ -21,8 +21,8 @@ func (s *partyServer) UpdateParty(c context.Context, req *pg.UpdatePartyRequest)
 		Id:        req.PartyId,
 		UserId:    req.RequesterId,
 		Title:     req.Title,
-		Lat:       float64(req.Lat),
-		Long:      float64(req.Long),
+		Lat:       req.Lat,
+		Long:      req.Long,
 		StartDate: start,
 	}
